refactor(filid): promote pr-summary to sub-fractal (T1-E)
Split core/utils/pr-summary-generator.ts (353 lines, 12 functions) into a
new core/pr-summary/ sub-fractal with organ-first decomposition:

- parsers/      (2 files) — parse-structure-check, parse-fix-requests
- aggregators/  (5 files) — collect, resolve, compute, map
- renderers/    (2 files) — render-markdown, severity-emoji
- pr-summary.ts — eponymous facade (generateHumanSummary orchestrator)
- INTENT.md     — 43 lines, 3-tier boundaries

Consumer paths updated: src/core/index.ts re-export + 2 test files + 1 MCP
tool format path. src/index.ts public surface unchanged.

- src/index.ts diff: empty (backward compat preserved)
- bundle delta: mcp-server.cjs +0.091% (<5% gate), all .mjs 0%
- tests: 846/850 pass (4 pre-existing context-injector env failures)
