+++ packages/filid/src/core/pr-summary/index.ts
@@ -1,0 +1,2 @@
+export * from './pr-summary.js';
+export type * from './pr-summary.js';
