+++ packages/filid/src/core/pr-summary/renderers/severity-emoji.ts
@@ -1,0 +1,14 @@
+import type { SummaryItemSeverity } from '../../../types/summary.js';
+
+export function severityEmoji(severity: SummaryItemSeverity): string {
+  switch (severity) {
+    case 'critical':
+      return '🚨';
+    case 'warning':
+      return '⚠️';
+    case 'info':
+      return 'ℹ️';
+    case 'pass':
+      return '✅';
+  }
+}
